fix test_vuln
