+++ api/app/tests/requests/test_vulns.py
@@ -1,0 +1,139 @@
+from uuid import UUID, uuid4
+
+import pytest
+from fastapi.testclient import TestClient
+from sqlalchemy.orm import Session
+
+from app.main import app
+from app.tests.medium.constants import (
+    USER1,
+)
+from app.tests.medium.utils import (
+    create_user,
+    headers,
+)
+
+client = TestClient(app)
+
+
+class TestUpdateVuln:
+    @pytest.fixture(scope="function", autouse=True)
+    def common_setup(self, testdb: Session):
+        self.user1 = create_user(USER1)
+        self.request1 = {
+            "title": "Example vuln",
+            "cve_id": "CVE-0000-0001",
+            "detail": "This vuln is example.",
+            "exploitation": "active",
+            "automatable": "yes",
+            "cvss_v3_score": 7.8,
+            "vulnerable_packages": [
+                {
+                    "name": "example-lib",
+                    "ecosystem": "pypi",
+                    "affected_versions": ["<2.0.0"],
+                    "fixed_versions": ["2.0.0"],
+                }
+            ],
+        }
+
+    def test_raise_400_if_given_vuln_id_is_default_vuln_id(self):
+        # Given
+        default_vuln_id = UUID(int=0)
+
+        # When
+        response = client.put(
+            f"/vulns/{default_vuln_id}", headers=headers(USER1), json=self.request1
+        )
+
+        # Then
+        assert response.status_code == 400
+        assert response.json()["detail"] == "Cannot create default vuln"
+
+    def test_raise_400_if_requested_title_is_None(self):
+        # Given
+        new_vuln_id = uuid4()
+        title_none_request = {
+            "cve_id": "CVE-0000-0001",
+            "detail": "This vuln is example.",
+            "exploitation": "active",
+            "automatable": "yes",
+            "cvss_v3_score": 7.8,
+            "vulnerable_packages": [
+                {
+                    "name": "example-lib",
+                    "ecosystem": "pypi",
+                    "affected_versions": ["<2.0.0"],
+                    "fixed_versions": ["2.0.0"],
+                }
+            ],
+        }
+        # When
+        response = client.put(
+            f"/vulns/{new_vuln_id}", headers=headers(USER1), json=title_none_request
+        )
+
+        # Then
+        assert response.status_code == 400
+        assert (
+            response.json()["detail"]
+            == "Both 'title' and 'detail' are required when creating a vuln."
+        )
+
+    def test_raise_400_if_requested_detail_is_None(self):
+        # Given
+        new_vuln_id = uuid4()
+        detail_none_request = {
+            "title": "Example vuln",
+            "cve_id": "CVE-0000-0001",
+            "exploitation": "active",
+            "automatable": "yes",
+            "cvss_v3_score": 7.8,
+            "vulnerable_packages": [
+                {
+                    "name": "example-lib",
+                    "ecosystem": "pypi",
+                    "affected_versions": ["<2.0.0"],
+                    "fixed_versions": ["2.0.0"],
+                }
+            ],
+        }
+        # When
+        response = client.put(
+            f"/vulns/{new_vuln_id}", headers=headers(USER1), json=detail_none_request
+        )
+
+        # Then
+        assert response.status_code == 400
+        assert (
+            response.json()["detail"]
+            == "Both 'title' and 'detail' are required when creating a vuln."
+        )
+
+    def test_raise_400_if_cvss_v3_score_is_out_of_range(self):
+        # Given
+        new_vuln_id = uuid4()
+        out_of_range_cvss_request = {
+            "title": "Example vuln",
+            "cve_id": "CVE-0000-0001",
+            "detail": "This vuln is example.",
+            "exploitation": "active",
+            "automatable": "yes",
+            "cvss_v3_score": 9999.9999,
+            "vulnerable_packages": [
+                {
+                    "name": "example-lib",
+                    "ecosystem": "pypi",
+                    "affected_versions": ["<2.0.0"],
+                    "fixed_versions": ["2.0.0"],
+                }
+            ],
+        }
+        # When
+        response = client.put(
+            f"/vulns/{new_vuln_id}", headers=headers(USER1), json=out_of_range_cvss_request
+        )
+
+        # Then
+        assert response.status_code == 400
+        assert response.json()["detail"] == "cvss_v3_score is out of range"
